TST: Move to skip based on tf installation
