--- pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-markers =
-    keras_toolkit: mark a test as an element of the keras_toolkit
